requirements.txt
openenv-core
fastapi
pydantic>=2.0.0
uvicorn
pinecone
openai
pytest
httpx
python-dotenv
